requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.3
daphne==4.0.0
Django==4.1.3
django-anymail==8.6
django-celery-results==2.4.0
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
hiredis==2.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.2.4
phonenumbers==8.13.0
prompt-toolkit==3.0.36
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==22.1.0
python-usernames==0.3.1
pytz==2022.6
redis==4.4.0
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.5.2
